refactor: inline model provider request
